site regarding views saperated
