+++ gnowsys-ndf/gnowsys_ndf/ndf/views/home.py
@@ -195,55 +195,3 @@ def help_page_view(request,page_name):
                                         },
                                         context_instance=RequestContext(request)
                                     )
-@get_execution_time
-def site_about(request):
-    # page_obj = Node.get_node_by_id(page_id)
-    return render_to_response(
-                                        "ndf/site_about.html",
-                                        {
-                                            'title': 'About',
-                                        },
-                                        context_instance=RequestContext(request)
-                                    )
-@get_execution_time
-def site_credits(request):
-    # page_obj = Node.get_node_by_id(page_id)
-    return render_to_response(
-                                        "ndf/site_credits.html",
-                                        {
-                                            'title': 'Credits',
-                                        },
-                                        context_instance=RequestContext(request)
-                                    )
-
-@get_execution_time
-def site_contact(request):
-    # page_obj = Node.get_node_by_id(page_id)
-    return render_to_response(
-                                        "ndf/site_contact.html",
-                                        {
-                                            'title': 'Contact',
-                                        },
-                                        context_instance=RequestContext(request)
-                                    )
-
-@get_execution_time
-def site_termsofuse(request):
-    # page_obj = Node.get_node_by_id(page_id)
-    return render_to_response(
-                                        "ndf/site_termsofuse.html",
-                                        {
-                                            'title': 'Terms Of Use',
-                                        },
-                                        context_instance=RequestContext(request)
-                                    )
-@get_execution_time
-def site_privacypolicy(request):
-    # page_obj = Node.get_node_by_id(page_id)
-    return render_to_response(
-                                        "ndf/site_privacypolicy.html",
-                                        {
-                                            'title': 'Privacy Policy',
-                                        },
-                                        context_instance=RequestContext(request)
-                                    )
